requirements.txt
asgiref==3.8.1
babel==2.16.0
Django==5.1
django-money==3.5.3
py-moneyed==3.0
setuptools==72.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
psycopg2-binary
django-timezone-field
pytz